requirements.txt
flask
pillow
cairosvg
requests